install-crate: move install action to new `cargo-gpu-install` crate
