+++ Cargo.toml
@@ -1,6 +1,7 @@
 [workspace]
 members = [
 	"crates/cargo-gpu",
+	"crates/cargo-gpu-install",
 	"crates/xtask",
 ]
 
@@ -24,6 +25,7 @@ license = "MIT OR Apache-2.0"
 
 [workspace.dependencies]
 spirv-builder = { git = "https://github.com/Rust-GPU/rust-gpu", rev = "a30bd43db45f2bfe260051f44141e5eaffcbb4b0", default-features = false }
+cargo-gpu-install = { path = "./crates/cargo-gpu-install" }
 anyhow = "1.0.98"
 clap = { version = "4.5.41", features = ["derive"] }
 crossterm = "0.29.0"
